+++ CHANGELOG.md
@@ -3431,3 +3431,4 @@
 - [2026-08-06 21:13] fix(test): prevent double tap duplicate action in test
 - [2026-08-06 21:24] perf(network): apply RepaintBoundary on animated network widgets
 - [2026-08-06 21:30] style(forms): adjust design tokens, typography, and contrast in forms
+- [2026-08-06 21:45] test(forms): add unit tests for forms business logic